@@ -1,6 +1,6 @@
 {
   "name": "custom-form-app",
-  "version": "1.1.0",
+  "version": "3.0.1",
   "scripts": {
     "tcli": "npm install gulp-cli -g && npm install -g @tibco-tcstk/cloud-cli && tcli",
     "ng": "ng",
@@ -16,42 +16,42 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^10.2.4",
-    "@angular/cdk": "^8.2.3",
-    "@angular/common": "^10.2.4",
-    "@angular/compiler": "^10.2.4",
-    "@angular/core": "^10.2.4",
-    "@angular/elements": "^10.2.4",
-    "@angular/flex-layout": "^10.0.0-beta.32",
-    "@angular/forms": "^10.2.4",
-    "@angular/material": "^8.1.1",
-    "@angular/material-moment-adapter": "^10.2.7",
-    "@angular/platform-browser": "^10.2.4",
-    "@angular/platform-browser-dynamic": "^10.2.4",
-    "@angular/router": "^10.2.4",
+    "@angular/animations": "^12.2.1",
+    "@angular/cdk": "^12.2.1",
+    "@angular/common": "^12.2.1",
+    "@angular/compiler": "^12.2.1",
+    "@angular/core": "^12.2.1",
+    "@angular/elements": "^12.2.1",
+    "@angular/flex-layout": "^12.0.0-beta.34",
+    "@angular/forms": "^12.2.1",
+    "@angular/material": "^12.2.1",
+    "@angular/material-moment-adapter": "^12.2.1",
+    "@angular/platform-browser": "^12.2.1",
+    "@angular/platform-browser-dynamic": "^12.2.1",
+    "@angular/router": "^12.2.1",
     "concat": "^1.0.3",
-    "core-js": "^3.8.3",
+    "core-js": "^3.16.2",
     "document-register-element": "^1.14.10",
     "fs-extra": "^8.1.0",
-    "hammerjs": "^2.0.8",
     "moment": "^2.29.1",
     "rxjs": "^6.6.3",
-    "tslib": "^2.1.0",
-    "zone.js": "~0.10.2"
+    "tslib": "^2.3.1",
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-builders/custom-webpack": "^10.0.1",
-    "@angular-devkit/build-angular": "^0.1100.7",
-    "@angular/cli": "^10.2.1",
-    "@angular/compiler-cli": "^10.2.4",
-    "@angular/language-service": "^10.2.4",
-    "@tibco-tcstk/component-template": "^2.2.3",
-    "@types/node": "^12.19.14",
-    "@webcomponents/webcomponentsjs": "^2.5.0",
+    "@angular-builders/custom-webpack": "^12.1.1",
+    "@angular-devkit/build-angular": "^12.2.1",
+    "@angular/cli": "^12.2.1",
+    "@angular/compiler-cli": "^12.2.1",
+    "@angular/language-service": "^12.2.1",
+    "@tibco-tcstk/component-template": "^3.0.1",
+    "@types/node": "^14.17.9",
+    "@webcomponents/webcomponentsjs": "^2.6.0",
     "child_process": "^1.0.2",
-    "codelyzer": "^5.1.2",
+    "codelyzer": "^6.0.0",
     "ts-node": "^8.10.2",
-    "tslint": "~6.1.0",
-    "typescript": "~4.0.5"
+    "tslint": "~6.1.3",
+    "typescript": "^4.3.5",
+    "webpack-dev-server": "^3.11.2"
   }
 }
